Added new CancelException class to handle when user wants to cancel a command. Refactored parser to clean it up a little.
