+++ src/main/java/seedu/duke/CancelException.java
@@ -1,0 +1,4 @@
+package seedu.duke;
+
+public class CancelException extends Exception {
+}
